r102
cache redis git-svn-id: https://torrentpier2.googlecode.com/svn/trunk@102 a8ac35ab-4ca4-ca47-4c2d-a49a94f06293

+++ upload/includes/cron/jobs/captcha_gen_gc.php
@@ -53,8 +53,10 @@ $del_ids = DB()->fetch_rowset("SELECT cap_id FROM ". BB_CAPTCHA ." WHERE cap_id 
 foreach ($del_ids as $del_id)
 {
 	$cap_img_path = CAPTCHA()->get_img_path(abs($del_id));
-	unlink($cap_img_path);
-
+	if(@fopen($cap_img_path, 'r'))
+	{
+		unlink($cap_img_path);
+    }
 	DB()->query("DELETE FROM ". BB_CAPTCHA ." WHERE cap_id = $del_id LIMIT 1");
 }
